# Fix and speedup utils.symmetric_difference()