--- a/tests/test_utils.py
+++ b/tests/test_utils.py
@@ -2251,7 +2251,15 @@ def test_to_segmented_index(obj, allow_nat, files_duration, root, expected):
                 audformat.filewise_index(['f1', 'f2']),
                 audformat.filewise_index(['f2', 'f3']),
             ],
-            audformat.filewise_index('f3'),
+            audformat.filewise_index(['f2', 'f3']),
+        ),
+        (
+            [
+                audformat.filewise_index(['f2', 'f3']),
+                audformat.filewise_index(['f1', 'f2']),
+                audformat.filewise_index(['f1', 'f2']),
+            ],
+            audformat.filewise_index(['f3', 'f2']),
         ),
         (
             [
@@ -2294,7 +2302,7 @@ def test_to_segmented_index(obj, allow_nat, files_duration, root, expected):
                 audformat.segmented_index(['f2', 'f1'], [0, 0], [1, 1]),
                 audformat.segmented_index(['f2', 'f3'], [0, 0], [1, 1]),
             ],
-            audformat.segmented_index('f3', 0, 1),
+            audformat.segmented_index(['f2', 'f3'], [0, 0], [1, 1]),
         ),
         (
             [
@@ -2302,11 +2310,7 @@ def test_to_segmented_index(obj, allow_nat, files_duration, root, expected):
                 audformat.segmented_index(['f2', 'f1'], [0, 0], [1, 1]),
                 audformat.segmented_index(['f2', 'f3'], [1, 1], [2, 2]),
             ],
-            audformat.segmented_index(
-                ['f2', 'f3'],
-                [1, 1],
-                [2, 2],
-            ),
+            audformat.segmented_index(['f2', 'f3'], [1, 1], [2, 2]),
         ),
         (
             [
@@ -2333,6 +2337,18 @@ def test_to_segmented_index(obj, allow_nat, files_duration, root, expected):
             [
                 audformat.segmented_index(['f1', 'f2'], [0, 0], [1, 1]),
                 audformat.segmented_index(['f2', 'f3'], [0, 0], [1, 1]),
+                audformat.filewise_index(['f1', 'f2']),
+            ],
+            audformat.segmented_index(
+                ['f1', 'f3', 'f1', 'f2'],
+                [0, 0, 0, 0],
+                [1, 1, pd.NaT, pd.NaT],
+            ),
+        ),
+        (
+            [
+                audformat.segmented_index(['f1', 'f2'], [0, 0], [1, pd.NaT]),
+                audformat.segmented_index(['f2', 'f3'], [0, 0], [pd.NaT, 1]),
                 audformat.filewise_index(['f1', 'f2']),
             ],
             audformat.segmented_index(