Closes #245

This fixes the results and tests for `audformat.utils.symmetric_difference()` for more than two objects,
and speeds up the implementation by basing it on `list` and `collections.Counter`.
It has the same speed as using `set`, but has the advantage of preserving the order.

**Benchmark results (code below)**

| Table type | Index samples | v0.14.3 | master | pull request |
| --- | --- | --- | --- | --- | 
| filewise | (100000, 500000, 100000) | - | 0.54 s  | 0.41 s |
| filewise | (100000, 500000, 10000) | - | 0.41 s  | 0.36 s |
| segmented | (100000, 500000, 100000) | - | 22.39 s | 8.55 s |
| segmented | (100000, 500000, 10000) | - | 19.74 s | 8.03 s |

![image](https://user-images.githubusercontent.com/173624/182375756-93e0c9b0-17cb-4218-9969-90e459466ddd.png)

---

Code for benchmark:

```python
import audformat
import numpy as np
import pandas as pd
import random
import time


random.seed(0)
samples = 100000


def main():
    nsamples = 100000
    repetitions = 10

    for description, samples in [
            (
                'same length',
                [nsamples, 5 * nsamples, nsamples],
            ),
            (
                '1/10 length',
                [nsamples, 5 * nsamples, int(nsamples / 10)],
            ),
    ]:  
        for index in [filewise, segmented]:
            times = []
            for _ in range(repetitions):
                objs = index(samples)
                times.append(measure(objs))
            print(
                f'Execution time {index.__name__} {description}: '
                f'{np.mean(times):.2f}s'
            )


def measure(objs):
    start = time.time()
    _ = audformat.utils.symmeric_difference(objs)
    end = time.time()
    return end - start


def filewise(samples):
    index1 = audformat.filewise_index([f"f{n}" for n in range(samples[0])])
    index2 = audformat.filewise_index([f"f{n}" for n in range(samples[1])])
    idx = [n for n in range(len(index2))]
    index3 = index2[random.sample(idx, k=samples[2])]
    return [index1, index2, index3]


def segmented(samples):
    index1 = audformat.segmented_index(
        [f"f{n}" for n in range(samples[0])],
        random.choices([0, 1, 2, 3], k=samples[0]),
        random.choices([4, 5, 6, pd.NaT], k=samples[0]),
    )
    index2 = audformat.segmented_index(
        [f"f{n}" for n in range(samples[1])],
        random.choices([0, 1, 2, 3], k=samples[1]),
        random.choices([4, 5, 6, pd.NaT], k=samples[1]),
    )
    idx = [n for n in range(len(index2))]
    index3 = index2[random.sample(idx, k=samples[2])]
    return [index1, index2, index3]


if __name__ == '__main__':
    main()
```