@frankenjoe 2022-08-02
I propose we close this PR as `difference()` with two objects covers `symmetric_difference()` and with more than two objects `symmetric_difference()` is not really useful.