@hagenw 2022-08-03
Agree, closing in favor of https://github.com/audeering/audformat/pull/264